backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.6.1
pytest==7.4.3
pytest-asyncio==0.23.2
pillow==10.2.0
